Changed to put everything in glfw3.jl
