+++ .gitignore
@@ -1,4 +1,3 @@
 deps/*
 !deps/build.jl
 
-*.txt
